Fixes misleading contramap example
The current version in the documentation old works because the values
are incorrectly transposed and are of the same type. This update calls
out that the case class needs to be ordered the same way for the proper
Iso instance to be derived.

It seems like exposing something along of the lines of Hlist's `rotate`
functions might be useful.