Fixes memory leak in prepared statement cache.  Fixes #1143
Prepared statements were created on the server and cached as part of the
"prepare cache."  However, despite entries being removed from the
in-memory cache, the evicted entries were not removed from Postgres
itself.

This modifies SemispaceCache to keep track of evicted elements.  When a
Session is recycled, we close any prepared statements that have been
evicted.